Update rules.md
I find the last line of explanation a bit vague in Documentation → The Rules → [Anatomy of a typical Harp application](http://harpjs.com/docs/development/rules):

````
+- hello-world.jade  <-- must have an index.html or index.jade file
````

This line shows a custom-named html/jade file (`hello-world.jade`) while explaining that there must be at least one `index` .html or .jade file (of which none exist in the example's `articles/` directory).

Maybe it's too early in the morning for me, but to me this appears somewhat non self-explanatory. Is a fixed-name (`index`.html|jade) file required or not? Or is it required that there's *at least* one jade or html file in here? In other words, what's actually required?

Adding to the confusion: there's an `index.jade` file in the `public/` dir, which I believe *has* to be explicitly named `index`, right?

Perhaps something like this might eliminate any confusion?

````
+- hello-world.jade  <-- must have at least one *.html or *.jade file
````