Update the Heroku deployment docs
zeke's harp buildpack has been deprecated, and in the meantime it's now
become simpler to use the official Heroku nodejs buildpack directly.

If no Procfile is provided, Heroku runs `npm start` by default, meaning
that in theory the creation of the Procfile could be skipped - and
instead a "start" script entry added to `package.json`. However
on Windows the `--port ${PORT:-9000}` parameter isn't expanded by
`npm start`, so the "start" script would have to omit it.

As such, it's just easier to add a Procfile directly (which can specify
the Heroku-specific $PORT variable) and leave it up to the user what
they add for a "start" script in package.json. For people used to Heroku
it's also more explicit what command the web process is running, if they
see a Procfile in the repo.

Fixes #53.